add user list command
- Commands will be based on an <object> <action> model because people
think in terms of objects; it will be easier to organize / discover what
can be done based on the <object> <action> sequence.
- Create a UI layer to wrap mutex locks and table printer. This is so
the display logic is split from the command logic.
- Use real UI with fake buffers for tests so that the raw output of the
UI can be tested instead of each individual call to UI. This should make
the tests less brittle when making UI/UX changes.
- Use landoop/tableprinter because why reinvent the wheel and they have
a json mode for when the product adds that feature.
- Apparently viper does not have the ability to initialize a config
despite what the documentation says. So logic needed to be added to
create config dir and file for first time users.